requirements.txt
wakeonlan>=3.0.0
django-crontab>=0.7.1
subprocess
datetime